+++ ttnn/cpp/ttnn/operations/experimental/deepseek/moe/moe_gate_mm/device/kernels/top8_merge_sfpu.h
@@ -436,7 +436,8 @@ inline void _llk_math_top8_merge_init_() {
 
 template <uint32_t column_idx>
 inline void _llk_math_top8_merge_() {
-    _llk_math_eltwise_unary_sfpu_params_(ckernel::sfpu::_top8_merge_<column_idx>, 0, VectorMode::RC_custom);
+    _llk_math_eltwise_unary_sfpu_params_(
+        ckernel::sfpu::_top8_merge_<column_idx>, 0, static_cast<int>(VectorMode::RC_custom));
 }
 
 #endif
