+++ ttnn/cpp/ttnn/operations/reduction/accumulation/ema/kernels/compute/ema_compute.cpp
@@ -60,7 +60,7 @@ inline void ema_sfpi_tile(
     bool first_sample) {
     MATH(_llk_math_eltwise_binary_sfpu_params_(
         ema_sfpi_face, inp_dst_index, prv_dst_index, out_dst_index,
-        VectorMode::RC, alpha, beta, first_sample));
+        static_cast<int>(VectorMode::RC), alpha, beta, first_sample));
 }
 // ------------------------------------------------------------------------------------------------
 */
